Moving exceptions to jobs
