+++ tests/jobs/test_Jobs.py
@@ -1,6 +1,6 @@
 import pytest
 from edsl.agents import Agent
-from edsl.exceptions.jobs import JobsRunError
+from edsl.jobs.exceptions import JobsRunError
 from edsl.agents.exceptions import AgentCombinationError
 from edsl.interviews import Interview
 from edsl.jobs import Jobs
